pyproject.toml
[build-system]
requires = ["hatchling>=0.25.1", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "mailboxer-python"
description = "Client library for Mailboxer"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "BSD 3-Clause License" }

classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = ["requests", "URLObject"]
dynamic = ["version"]

authors = [{ name = "Rotem Yaari", email = "vmalloc@gmail.com" }]

[project.urls]
"Homepage" = "https://github.com/getslash/mailboxer-python"

[project.optional-dependencies]
testing = ["dataclasses_json", "Flask-Loopback", "pylint", "pytest"]

[tool.hatch.build.targets.wheel]
packages = ["mailboxer"]

[tool.hatch.version]
source = "vcs"
